- Can't specify a static channel number for the server, it'll crash This version at least starts on the phone.
git-svn-id: file:///srv/svn/openuat-filtered-without-thirdparty/trunk@807 01b9256e-9321-419f-88ee-03bd83f56c6c

+++ src/org/openuat/util/BluetoothRFCOMMServer.java
@@ -69,7 +69,7 @@ public class BluetoothRFCOMMServer extends HostServerBase {
 
 		// construct the Bluetooth service URL
 		serviceURL = "btspp://localhost:" + serviceUUID + 
-			(channel != null ? ":" + channel : "") + ";name=" + serviceName + 
+			(channel != null ? ":" + channel.intValue() : "") + ";name=" + serviceName + 
 			";authenticate=false;encrypt=false;authorize=false;master=false";
 		// the service itself will be created later when calling startListening
 	}
